Move some proposal functions to internal
